refactor: Move homepage files to app
